CIGARStrings.jl

CIGARStrings.jl provide functionality for parsing and working with Concise Idiosyncratic Gapped Alignment Report - or CIGAR - strings. CIGARs were popularized by the SAM format, and are a compact run length encoding notation to represent pairwise alignments. They can be found in the SAM, BAM, PAF, and GFA formats.

For example, the following pairwise alignment of a query to a reference:

    Q: TAGAT---TAGCTAC
       ||||    ||  | |
    R: TAGAACCATA--TGC

Can be represented by the CIGAR 5M3D2M2I3M, representing:

5 matches/mismatches

Then, 3 deletions

Then, 2 matches/mismatches

Then, 2 insertions

Finally, 3 matches/mismatches.

A CIGAR strings is always written in terms of the query, and not the reference.

Individual alignment operations

One run of identical alignment operations, e.g. "5 matches/mismatches" are represented by a single CIGARElement. Conceptually, a CIGARElement is an alignment operation (represented by a CIGAROp) and a length:

CIGARStrings.CIGARElement — Type

CIGARElement(op::CIGAROp, len::Integer)

Type representing a single element in a CIGAR string, consisting of a CIGAROp and a length. For example, in 15S198M1D15M, the four elements are 15S, 198M, 1D, and 15M. Access the operation and the length with the properties .op and .len. Note that currently, the largest supported length is 268435455. Operations cannot have length zero.

See also: CIGAR, CIGAROp

Examples

julia> e = CIGARElement(OP_X, 3)
CIGARElement(OP_X, 3)

julia> e.len
3

julia> e.op
OP_X

CIGARStrings.CIGAROp — Type

CIGAROp

1-byte primitive type representing a CIGAR operation. Module-level constants for each value of this type are exported, e.g. OP_M.

A CIGAROp is guaranteed to be a 1-byte primitive with the same bitwise representation as a UInt8 with the value given as N in the table below, e.g. reinterpret(UInt8, OP_I) === 0x01.

The Consumes entry below signifies if the operation advances the position in the query (Q), the reference (R) and the alignment (A). E.g. if the CIGARElement 9D begins at query, ref, aln positions (A, B, C), then the positions are (A, B+9, C+9) after.

N	Consumes	Variable	Char	Description

0	Q R A	OP_M	M	Alignment match or mismatch

1	Q A	OP_I	I	Insertion relative to the reference

2	R A	OP_D	D	Deletion relative to the reference

3	R	OP_N	N	Reference skipped from the alignment (usually an intron)

4	Q	OP_S	S	Soft clip (semantically identical to hard clip)

5	Q	OP_H	H	Hard clip (semantically identical to soft clip)

6		OP_P	P	Padding, position not present in query or reference

7	Q R A	OP_Eq	=	Alignment match, not mismatch

8	Q R A	OP_X	X	Alignment mismatch

See also: CIGARElement

Extended help

M means a match or a mismatch. By default, most programs will emit M instead of X or =, since the important part of the CIGAR is where the insertions and deletions are placed. To determine which bases in an M are matches and mismatches, the two sequences can be compared, base-wise.

N means that the region is spans an intron, which means the query sequence is deleted, but not due to an actual deletion (which would be a D operation). It can also be used for other uses where the reference bases is missing for another reason than a deletion, if such a use case is found.

S and H are semantically identical. They both signify the end(s) of the query sequence which are not part of the alignment. They differ only in whether the clipped bases are written in the SEQ field of a SAM record. Typically, hard-clipped bases are present as soft clipped bases in another record, such that writing them in both records would be wasteful.

P is only used for a multiple sequence alignment. If a third sequence contains an insertion relative to both the query and the reference, the query has a P at this position, indicating it's present in neither the query nor reference.

Most alignments only contain the operations M, I, D, S.

CIGARs

A CIGAR string is represented by an AbstractCIGAR, which currently has two subtypes: CIGAR and BAMCIGAR. These types differ in their memory layout: The former stores the CIGAR as its ASCII representation (as used in the SAM format), and the latter stores it in a binary format (as used in the BAM format). Both typs store its underlying data as an ImmutableMemoryView{UInt8}.

CIGARStrings.AbstractCIGAR — Type

abstract type AbstractCIGAR

This abstract type is (not yet) a defined interface. Its concrete subtypes are CIGAR and BAMCIGAR.

The API for these two types are almost interchangeable, so examples below will use CIGAR, since its plaintext representation makes examples easier. See BAMCIGAR section for a list of all differences between the two types.

CIGAR strings are validated upon construction

julia> CIGAR("2M1D3M")
CIGAR("2M1D3M")

julia> CIGAR("1M1W1S")
ERROR: Error around byte 4: Invalid operation. Possible values are "MIDNSHP=X".
[...]

Since CIGAR strings occur in various bioinformatics file formats, it is expected that users of CIGARStrings.jl will construct CIGARs from a view into a buffer storing a chunk of the file.

This is zero-copy, and ought not to allocate (although it currently does, due to a limitation of the Julia compiler):

For example:

julia> data = b"some format with CIGAR: 15M9D18M etc";

julia> buffer = collect(data);

julia> c = CIGAR(view(buffer, 25:32))
CIGAR("15M9D18M")

CIGARStrings.CIGAR — Type

CIGAR <: AbstractCIGAR

A CIGAR string represents the sequence of insertions, matches and deletions that comprise a pairwise alignment. Construct a CIGAR from any object x where MemoryView(x) returns a MemoryView{UInt8}, i.e. any memory-backed bytearray, or string.

Use CIGARStrings.try_parse to attempt to parse a CIGAR string without throwing an exception if the data is invalid.

See also: CIGARElement

Extended help

CIGAR strings are sequences of CIGARElement, from the 5' to the 3' of the query (or N- to C-terminal for amino acids). CIGAR strings comprise the entire query, i.e. the sum of lengths of elements with the XMI=SH operations equals the length of the query.

For example, the query AGCGTAGCACACC that aligns from query base 5 and ref base 1002, like this:

Q: 5    TAG--CACACC   13
R: 1002 TAGGACAC-CC 1011

Is summarized by the CIGAR 4S3M2D3M1I2M. The operations HX=PN are more rarely used, see CIGAROp for a description of the operations.

CIGARs are iterable, and returns its CIGARElements, in order:

julia> collect(CIGAR("2M1D3M"))
3-element Vector{CIGARElement}:
 CIGARElement(OP_M, 2)
 CIGARElement(OP_D, 1)
 CIGARElement(OP_M, 3)

They can be converted back to strings using string(::CIGAR), or printed into an IO, in which case their normal string representation is recovered:

julia> c = CIGAR("2M1D3M");

julia> string(c)
"2M1D3M"

julia> io = IOBuffer(); print(io, c); String(take!(io))
"2M1D3M"

The memory underlying the CIGAR types can be obtained with MemoryView(x) using the MemoryViews.jl package:

julia> c = CIGAR("19M");

julia> println(MemoryView(c))
UInt8[0x31, 0x39, 0x4d]

julia> println(MemoryView(BAMCIGAR(c)))
UInt8[0x30, 0x01, 0x00, 0x00]

Basic information about a CIGAR

The reference, query and alignment length can be obtained with the functions ref_length, query_length and aln_length.

In the alignment below, represented as 5M3D2M2I3M, the query length is 12, since there are 12 query symbols, the reference length is 13, and the alignment length is 15.

    Q: TAGAT---TAGCTAC
       ||||    ||  | |
    R: TAGAACCATA--TGC

We always have aln_length(c) ≥ max(query_length(c), ref_length(c))

julia> c = CIGAR("5M3D2M2I3M");

julia> ref_length(c)
13

julia> query_length(c)
12

julia> aln_length(c)
15

The alignment identity (number of matches, not mismatches divided by alignment length) can be obtained with aln_identity. Since the number of mismatches may not be known from the CIGAR itself (i.e. "100M" could have anywhere from 0 to 100 mismatches), the mismatches is passed in as an argument:

julia> aln_identity(CIGAR("3M1D3M1I2M"), 2)
0.6

Position translation

Sometimes it may be necessary to answer questions of the form "which reference position does query position 8 align to?".

As an example, consider the alignment below. The query position (QP), reference position (RP) and alignment position (AP) are also written in this alignment.

    QP:12345   6789012

    Q: TAGAT---TAGCTAC
       ||||    ||  | |
    R: TAGAACCATA--TGC

    RP:1234567890  123
    AP:123456789012345

We can see that query position 6 aligns to reference position 9, which is also alignment position 9.

These position translations can be obtained using the functions query_to_ref, query_to_aln, ref_to_query, ref_to_aln, aln_to_query and aln_to_ref. They return a Translation object that contains two properties: .pos and .kind.

When a position translation has a straightforward answer, the .kind property is CIGARStrings.pos, and the .pos field is the corresponding position:

julia> c = CIGAR("4M3D2M2I3M"); # alignment above

julia> query_to_ref(c, 6)
Translation(pos, 9)

julia> aln_to_query(c, 9)
Translation(pos, 6)

Note that these operations are in linear time, as they scan the CIGAR string from the beginning. If the CIGAR string is long, and many of these operations are done on the same string, it may be better to construct another data structure for the alignment, which allows logarithmic-time queries, e.g. the ones in BioAlignments.jl.

CIGARStrings.Translation — Type

Translation(kind::TranslationKind, pos::Int)

The result of translating from a position in the coordinate system of the query / reference / alignment to a position in one of the others. This type contains two documented properies: kind::TranslationKind and pos::Int.

If the resulting position is outside the target coordinate, .kind == outside and pos == 0

If the position maps to a non-gap symbol in the other coodinate system, .kind == pos, and the position is the pos'd symbol in the target coordinate system.

If the position maps to a gap, pos is the position of the symbol before the gap, and kind == gap. When translating to the aln coordinate system, the kind is never gap.

See also: CIGARStrings.TranslationKind

Examples:

julia> c = CIGAR("2M3D2M2I1M");

julia> for i in 0:8
           refpos = query_to_ref(c, i)
           println(refpos.pos, ' ', refpos.kind)
       end
0 outside
1 pos
2 pos
6 pos
7 pos
7 gap
7 gap
8 pos
0 outside

CIGARStrings.TranslationKind — Type

TranslationKind

This enum has values outside, pos and gap. It represents the result of translating a position between query, reference and alignment.

If outside, the input position translates to a position outside the alignment

If pos, the input position corresponds to a non-gap symbol in the alignment

If gap, the input position maps to a gap.

See also: Translation

Errors and error recovery

CIGARStrings.jl allows you to parse a poential CIGAR string without throwing an exception if the data is invalid, using the function CIGARStrings.try_parse.

CIGARStrings.CIGARError — Type

CIGARError

Exception kind thrown by the package CIGARStrings. CIGARErrors contain two properties: .kind, returning a CIGARErrorType, and .index, returning an Int, pointing to the approximate byte index where the exception was encountered.

The kind and pos are stable API, in the sense that e.g. an integer overflow at position 55 will always be represented by a CIGARError(55, CIGARStrings.Errors.IntegerOverflow). However, the same invalid CIGAR string may produce multiple different errors, and which error is produced in that case is NOT stable API, because that depends on specifics parsing internals.

julia> ce = CIGARStrings.try_parse(CIGAR, "15M9");

julia> ce.index
4

julia> ce.kind
Truncated::CIGARErrorType = 0x05

CIGARStrings.Errors.CIGARErrorType — Type

CIGARErrorType

Single-byte enum representing the different kind of errors returned or thrown by the package CIGARStrings. This currently implemented list of error kinds is not exhaustive, in that more could be added in minor versions of this package.

See also: CIGARError

CIGARStrings.try_parse — Function

try_parse(::Type{CIGAR}, x)::Union{CIGAR, CIGARError}

Cast x to a MemoryView{UInt8}, and try parsing a CIGAR from it. If the parsing is unsuccessful, return a CIGARError

Examples

julia> c = CIGARStrings.try_parse(CIGAR, "2S1M9I");

julia> c isa CIGAR # success
true

julia> c = CIGARStrings.try_parse(CIGAR, "1S7H9M1S");

julia> c.kind
InvalidHardClip::CIGARErrorType = 0x03

The BAMCIGAR type

A CIGAR in the BAM format is stored in an array of 32-bit integers. However, in order to make zero-copy CIGARs possible, the BAMCIGAR type is backed by an ImmutableMemoryView{UInt8} instead, with the same memory layout as its equivalent Memory{UInt32}.

CIGARStrings.BAMCIGAR — Type

BAMCIGAR <: AbstractCIGAR

A BAMCIGAR is an alternative representation of a CIGAR, stored compactly in 32-bit integers. Semantically, a BAMCIGAR behaves much similar to a CIGAR.

Construct a BAMCIGAR either from a CIGAR, taking an optional Vector{UInt8} to use as backing storage, or using CIGARStrings.try_parse.

Examples

julia> c = CIGAR("9S123M1=3I15M2H");

julia> b = BAMCIGAR(c, UInt8[]); # use existing storage

julia> c == b
true

julia> CIGAR(b)
CIGAR("9S123M1=3I15M2H")

A BAMCIGAR can be constructed from its binary representation, using any type which implements MemoryViews.MemoryView:

julia> BAMCIGAR("\x54\4\0\0\x70\4\0\0")
BAMCIGAR(CIGAR("69S71M"))

This is not zero-cost: Like CIGAR the type contains some metadata and is validated upon construction.

Like CIGAR, the try_parse function can be used:

julia> CIGARStrings.try_parse(BAMCIGAR, "\x5f\4\0\0\x70\4\0\0")
CIGARStrings.CIGARError(1, CIGARStrings.Errors.InvalidOperation)

CIGAR and BAMCIGAR can be converted ifallably to each other:

julia> c = CIGAR("6H19S18M1I22=8I2S");

julia> b = BAMCIGAR(c);

julia> b == c
true

julia> CIGAR(b) == c
true

Note that printing a BAMCIGAR allocates, because it needs to allocate a new piece of memory to store its ASCII representation. For high performance applications, the function cigar_view! may be used:

CIGARStrings.cigar_view! — Function

cigar_view!(v::Vector{UInt8}, x::BAMCIGAR)::ImmutableMemoryView{UInt8}

Write the ASCII (i.e. CIGAR) representation x into v, emptying v's original content. A memory view of v is returned:

Examples

julia> v = [0x01, 0x02, 0x03];

julia> bc = BAMCIGAR(CIGAR("151M3D20M"));

julia> mem_view = cigar_view!(v, bc);

julia> mem_view == v
true

julia> String(mem_view) == string(CIGAR(bc))
true