CIGARStrings.jl reference

CIGARStrings.CIGAR — Type

CIGAR <: AbstractCIGAR

A CIGAR string represents the sequence of insertions, matches and deletions that comprise a pairwise alignment. Construct a CIGAR from any object x where MemoryView(x) returns a MemoryView{UInt8}, i.e. any memory-backed bytearray, or string.

Use CIGARStrings.try_parse to attempt to parse a CIGAR string without throwing an exception if the data is invalid.

See also: CIGARElement

Extended help

CIGAR strings are sequences of CIGARElement, from the 5' to the 3' of the query (or N- to C-terminal for amino acids). CIGAR strings comprise the entire query, i.e. the sum of lengths of elements with the XMI=SH operations equals the length of the query.

For example, the query AGCGTAGCACACC that aligns from query base 5 and ref base 1002, like this:

Q: 5    TAG--CACACC   13
R: 1002 TAGGACAC-CC 1011

Is summarized by the CIGAR 4S3M2D3M1I2M. The operations HX=PN are more rarely used, see CIGAROp for a description of the operations.

CIGARStrings.CIGARElement — Type

CIGARElement(op::CIGAROp, len::Integer)

Type representing a single element in a CIGAR string, consisting of a CIGAROp and a length. For example, in 15S198M1D15M, the four elements are 15S, 198M, 1D, and 15M. Access the operation and the length with the properties .op and .len. Note that currently, the largest supported length is 268435455. Operations cannot have length zero.

See also: CIGAR, CIGAROp

Examples

julia> e = CIGARElement(OP_X, 3)
CIGARElement(OP_X, 3)

julia> e.len
3

julia> e.op
OP_X

CIGARStrings.CIGAROp — Type

CIGAROp

1-byte primitive type representing a CIGAR operation. Module-level constants for each value of this type are exported, e.g. OP_M.

A CIGAROp is guaranteed to be a 1-byte primitive with the same bitwise representation as a UInt8 with the value given as N in the table below, e.g. reinterpret(UInt8, OP_I) === 0x01.

The Consumes entry below signifies if the operation advances the position in the query (Q), the reference (R) and the alignment (A). E.g. if the CIGARElement 9D begins at query, ref, aln positions (A, B, C), then the positions are (A, B+9, C+9) after.

N	Consumes	Variable	Char	Description

0	Q R A	OP_M	M	Alignment match or mismatch

1	Q A	OP_I	I	Insertion relative to the reference

2	R A	OP_D	D	Deletion relative to the reference

3	R	OP_N	N	Reference skipped from the alignment (usually an intron)

4	Q	OP_S	S	Soft clip (semantically identical to hard clip)

5	Q	OP_H	H	Hard clip (semantically identical to soft clip)

6		OP_P	P	Padding, position not present in query or reference

7	Q R A	OP_Eq	=	Alignment match, not mismatch

8	Q R A	OP_X	X	Alignment mismatch

See also: CIGARElement

Extended help

M means a match or a mismatch. By default, most programs will emit M instead of X or =, since the important part of the CIGAR is where the insertions and deletions are placed. To determine which bases in an M are matches and mismatches, the two sequences can be compared, base-wise.

N means that the region is spans an intron, which means the query sequence is deleted, but not due to an actual deletion (which would be a D operation). It can also be used for other uses where the reference bases is missing for another reason than a deletion, if such a use case is found.

S and H are semantically identical. They both signify the end(s) of the query sequence which are not part of the alignment. They differ only in whether the clipped bases are written in the SEQ field of a SAM record. Typically, hard-clipped bases are present as soft clipped bases in another record, such that writing them in both records would be wasteful.

P is only used for a multiple sequence alignment. If a third sequence contains an insertion relative to both the query and the reference, the query has a P at this position, indicating it's present in neither the query nor reference.

Most alignments only contain the operations M, I, D, S.

CIGARStrings.Translation — Type

Translation(kind::TranslationKind, pos::Int)

The result of translating from a position in the coordinate system of the query / reference / alignment to a position in one of the others. This type contains two documented properies: kind::TranslationKind and pos::Int.

If the resulting position is outside the target coordinate, .kind == outside and pos == 0

If the position maps to a non-gap symbol in the other coodinate system, .kind == pos, and the position is the pos'd symbol in the target coordinate system.

If the position maps to a gap, pos is the position of the symbol before the gap, and kind == gap. When translating to the aln coordinate system, the kind is never gap.

See also: CIGARStrings.TranslationKind

Examples:

julia> c = CIGAR("2M3D2M2I1M");

julia> for i in 0:8
           refpos = query_to_ref(c, i)
           println(refpos.pos, ' ', refpos.kind)
       end
0 outside
1 pos
2 pos
6 pos
7 pos
7 gap
7 gap
8 pos
0 outside

CIGARStrings.TranslationKind — Type

TranslationKind

This enum has values outside, pos and gap. It represents the result of translating a position between query, reference and alignment.

If outside, the input position translates to a position outside the alignment

If pos, the input position corresponds to a non-gap symbol in the alignment

If gap, the input position maps to a gap.

See also: Translation

CIGARStrings.CIGARError — Type

CIGARError

Exception kind thrown by the package CIGARStrings. CIGARErrors contain two properties: .kind, returning a CIGARErrorType, and .index, returning an Int, pointing to the approximate byte index where the exception was encountered.

The kind and pos are stable API, in the sense that e.g. an integer overflow at position 55 will always be represented by a CIGARError(55, CIGARStrings.Errors.IntegerOverflow). However, the same invalid CIGAR string may produce multiple different errors, and which error is produced in that case is NOT stable API, because that depends on specifics parsing internals.

julia> ce = CIGARStrings.try_parse(CIGAR, "15M9");

julia> ce.index
4

julia> ce.kind
Truncated::CIGARErrorType = 0x05

CIGARStrings.Errors.CIGARErrorType — Type

CIGARErrorType

Single-byte enum representing the different kind of errors returned or thrown by the package CIGARStrings. This currently implemented list of error kinds is not exhaustive, in that more could be added in minor versions of this package.

See also: CIGARError

CIGARStrings.try_parse — Function

try_parse(::Type{CIGAR}, x)::Union{CIGAR, CIGARError}

Cast x to a MemoryView{UInt8}, and try parsing a CIGAR from it. If the parsing is unsuccessful, return a CIGARError

Examples

julia> c = CIGARStrings.try_parse(CIGAR, "2S1M9I");

julia> c isa CIGAR # success
true

julia> c = CIGARStrings.try_parse(CIGAR, "1S7H9M1S");

julia> c.kind
InvalidHardClip::CIGARErrorType = 0x03

CIGARStrings.OP_M — Constant

'M': Alignment match or mismatch

CIGARStrings.OP_I — Constant

'I': Insertion relative to the reference

CIGARStrings.OP_D — Constant

'D': Deletion relative to the reference

CIGARStrings.OP_S — Constant

'S': Soft clip (clipped sequence present in query)

CIGARStrings.OP_H — Constant

'H': Hard clip (clipped sequence not present in query)

CIGARStrings.OP_N — Constant

'N': Region skipped from the reference (usually an intron)

CIGARStrings.OP_P — Constant

'P': Padding

CIGARStrings.OP_X — Constant

'X': Alignment mismatch

CIGARStrings.OP_Eq — Constant

'=': Alignment match, not mismatch

CIGARStrings.PositionMapper — Type

PositionMapper{Params...}

Public struct returned by pos_to_pos(..., pos) when !isa(pos, Integer). This type may be referred to in stable code, but its API is defined by pos_to_pos

See also: pos_to_pos

CIGARStrings.query — Function

Coordinate system representing the 1-based index in the query sequence

CIGARStrings.ref — Function

Coordinate system representing the 1-based index in the reference sequence

CIGARStrings.aln — Function

Coordinate system representing the 1-based index in the alignment (i.e. query/ref including gaps)

CIGARStrings.ref_length — Function

ref_length(::AbstractCIGAR)::Int

Get the number of biosymbols in the reference of the CIGAR. This is the same as the lengths of all CIGARElements of type M, D, N, = and X.

See also: query_length, aln_length

Example

julia> ref_length(CIGAR("1S5M2D6M2I5M"))
18

CIGARStrings.query_length — Function

query_length(::AbstractCIGAR)::Int

Get the number of biosymbols in the query of the CIGAR. This is the same as the lengths of all CIGARElements of type M, I, S, H, = and X.

See also: ref_length, aln_length

Example

julia> query_length(CIGAR("1S5M2D6M2I5M"))
19

CIGARStrings.aln_length — Function

aln_length(::AbstractCIGAR)::Int

Get the number of biosymbols spanned by the alignment of the CIGAR. Clips, padding and skips are not part of the alignment, but still part of the CIGAR. Therefore, the alignment length is the same as the lengths of all CIGARElements of type M, I, D, = and X.

See also: query_length, ref_length

Example

julia> aln_length(CIGAR("1S5M2D6M2I5M"))
20

CIGARStrings.aln_identity — Function

aln_identity(::AbstractCIGAR, mismatches::Int)::Float64

Compute the alignment identity of the CIGAR, computed as the number of matches (not mismatches) divided by alignment length. Since the CIGAR itself may not provide information about the precise number of mismatches, the amount of mismatches is an argument. The result is always in [0.0, 1.0].

Example

julia> aln_identity(CIGAR("3M1D3M1I2M"), 2)
0.6

CIGARStrings.query_to_ref — Function

query_to_ref(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based reference position aligning to query position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> query_to_ref(c, 4)
Translation(pos, 6)

CIGARStrings.query_to_aln — Function

query_to_aln(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based alignment position aligning to query position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> query_to_aln(c, 8)
Translation(pos, 10)

CIGARStrings.ref_to_query — Function

ref_to_query(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based query position aligning to reference position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> ref_to_query(c, 7)
Translation(pos, 9)

CIGARStrings.ref_to_aln — Function

ref_to_aln(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based alignment position aligning to reference position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> ref_to_aln(c, 7)
Translation(pos, 11)

CIGARStrings.aln_to_query — Function

aln_to_query(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based query position aligning to alignment position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> aln_to_query(c, 10)
Translation(pos, 8)

CIGARStrings.aln_to_ref — Function

aln_to_ref(x::AbstractCIGAR, pos::Integer)::Int

Get the 1-based reference position aligning to alignment position pos. See Translation for more details.

Examples

julia> c = CIGAR("3M2D1M4I2M");

julia> aln_to_ref(c, 9)
Translation(gap, 6)